separate controller added for multiple simulation simulation, the multsim is disabled when not login
