+++ OpenRobertaParent/OpenRobertaServer/staticResources/index.html
@@ -92,7 +92,7 @@ limitations under the License.
                             <li><a href="#" id="menuRunSim" class="typcn typcn-simulation" lkey="Blockly.Msg.MENU_START_SIM">run in simulation</a></li>
                             <li><a href="#" id="menuNewProg" class="typcn typcn-document-add" lkey="Blockly.Msg.MENU_NEW">neu ...</a></li>
                             <li class="login"><a href="#" id="menuListProg" class="typcn typcn-th-list" lkey="Blockly.Msg.MENU_LIST_PROG">Liste ...</a></li>
-                            <li><a href="#" id="multipleSimNav" class="typecn " data-toggle="modal" data-target="#simModal">Run mult sim</a></li>
+                            <li class="login"><a href="#" id="multipleSimNav" class="typecn " data-toggle="modal" data-target="#simModal">Run mult sim</a></li>
                             <li><a href="#" id="menuListExamples" class="typcn typcn-th-list" lkey="Blockly.Msg.MENU_LIST_PROG_EXAMPLES">Beispielprogramme
                                     ...</a></li>
                             <li class="login"><a href="#" id="menuSaveProg" class="typcn typcn-cloud-storage" lkey="Blockly.Msg.MENU_SAVE">speichern</a></li>
