+++ OpenRobertaParent/OpenRobertaServer/staticResources/js/main.js
@@ -32,6 +32,7 @@ require.config({
         'logList.controller' : '../app/roberta/controller/logList.controller',
         'logList.model' : '../app/roberta/models/logList.model',
         'menu.controller' : '../app/roberta/controller/menu.controller',
+        'multSim.controller' : '../app/roberta/controller/multSim.controller',
         'progCode.controller' : '../app/roberta/controller/progCode.controller',
         'progDelete.controller' : '../app/roberta/controller/progDelete.controller',
         'progHelp.controller' : '../app/roberta/controller/progHelp.controller',
@@ -128,7 +129,7 @@ require.config({
 });
 
 require([ 'require', 'wrap', 'jquery', 'jquery-cookie', 'guiState.controller', 'progList.controller', 'logList.controller', 'confList.controller',
-        'progDelete.controller', 'confDelete.controller', 'progShare.controller', 'cookieDisclaimer.controller', 'menu.controller', 'user.controller',
+        'progDelete.controller', 'confDelete.controller', 'progShare.controller', 'cookieDisclaimer.controller', 'menu.controller','multSim.controller', 'user.controller',
         'robot.controller', 'program.controller', 'progSim.controller', 'progCode.controller', 'progDelete.controller', 'progHelp.controller',
         'progInfo.controller', 'progRun.controller', 'configuration.controller', 'language.controller', 'socket.controller', 'progTutorial.controller', 'volume-meter', 'user.model', 'webview.controller' ], function(
         require) {
@@ -144,6 +145,7 @@ require([ 'require', 'wrap', 'jquery', 'jquery-cookie', 'guiState.controller', '
     logListController = require('logList.controller');
     cookieDisclaimer = require('cookieDisclaimer.controller');
     menuController = require('menu.controller');
+    multSimController = require('multSim.controller');
     progDeleteController = require('progDelete.controller');
     progListController = require('progList.controller');
     galleryListController = require('galleryList.controller');
@@ -195,6 +197,7 @@ function init() {
         progRunController.init();
         menuController.init();
         tutorialController.init();
+        multSimController.init();
         
         cookieDisclaimer.init();
         $(".cover").fadeOut(100, function() {
